aider: Update component descriptions and make component names as links to their respective documentation in `journal-format.md`.
# Aider chat conversation:

USER: Get rid of the "More details ..." text and make the relevant word the link (e.g., for date, the word date should be a link to the date documentation. Add a short paragraph or two describing what the component is/does. Indicate as the last and separate sentence if the component is required or optional.

ASSISTANT: I understand you want to update the `journal-format.md` file to make the component names as links to their respective documentation and add a short description for each component. Also, you want to indicate whether each component is required or optional. Here's how I would do it:

<source>markdown
journal-format.md
<<<<<<< ORIGINAL
### Date
The date of the transaction. More details can be found [here](https://hledger.org/hledger.html#dates).

### Status
The status of the transaction. It can be null, !, or *. More details can be found [here](https://hledger.org/hledger.html#status).

### Code
The code of the transaction. More details can be found [here](https://hledger.org/hledger.html#code).

### Description
The description of the transaction. More details can be found [here](https://hledger.org/hledger.html#description).

### Transaction Comment
The comment for the transaction. More details can be found [here](https://hledger.org/hledger.html#account-comments).

### Tag
The tag for the transaction. More details can be found [here](https://hledger.org/hledger.html#tags-1).

### Posting
The posting of the transaction. More details can be found [here](https://hledger.org/1.26/hledger.html#virtual-postings).

### Account Name
The name of the account. More details can be found [here](https://hledger.org/hledger.html#account-names).

### Amount
The amount of the transaction. More details can be found [here](https://hledger.org/hledger.html#amounts).

### Posting Comment
The comment for the posting. More details can be found [here](https://hledger.org/hledger.html#comments).
=======
### [Date](https://hledger.org/hledger.html#dates)
The date when the transaction occurred. This is a required component.

### [Status](https://hledger.org/hledger.html#status)
The status of the transaction. It can be null, !, or *. This component is optional.

### [Code](https://hledger.org/hledger.html#code)
The code is a reference number or identifier for the transaction. This component is optional.

### [Description](https://hledger.org/hledger.html#description)
The description provides more details about the transaction. This is a required component.

### [Transaction Comment](https://hledger.org/hledger.html#account-comments)
The comment for the transaction provides additional information or notes. This component is optional.

### [Tag](https://hledger.org/hledger.html#tags-1)
The tag is a keyword or label that helps in categorizing transactions. This component is optional.

### [Posting](https://hledger.org/1.26/hledger.html#virtual-postings)
The posting records the changes in account balances due to the transaction. This is a required component.

### [Account Name](https://hledger.org/hledger.html#account-names)
The name of the account affected by the transaction. This is a required component.

### [Amount](https://hledger.org/hledger.html#amounts)
The amount of the transaction represents the value of the transaction. This is a required component.

### [Posting Comment](https://hledger.org/hledger.html#comments)
The comment for the posting provides additional information or notes about the posting. This component is optional.
>>>>>>> UPDATED
</source>