aider: Move bullet points to their respective sections and modify them into short paragraphs.
# Aider chat conversation:

USER: In the general information section, it seems to me those notes should go in their respective sections, but I'm not sure which should go where. Move those bullet points to their correct location and modify them so they are a short paragraph instead of a bullet point.

ASSISTANT: To move the bullet points to their respective sections, we need to identify where each bullet point belongs. Here's my understanding:

1. The first bullet point about indentation seems to be a general rule that applies to the entire file, so it might be best to keep it in the General Information section but rephrase it as a paragraph.

2. The second bullet point about secondary dates and posting dates belongs in the Date section.

3. The third bullet point about account comments ending at the first posting belongs in the Transaction Comment section.

Here are the changes I propose:

<source>markdown
journal-format.md
<<<<<<< ORIGINAL
## General Information

* \<indent\> must be a minimum of two spaces and is required if the line it's
  on is being used.

* ledgers [secondary dates](https://hledger.org/hledger.html#secondary-dates)
  are supported but not recommended, stick with [posting
  dates](https://hledger.org/hledger.html#posting-dates)

* Multiple account comment lines can be added, but they end at the first posting
=======
## General Information

Indentation must be a minimum of two spaces and is required if the line it's on is being used.

### [Date](https://hledger.org/hledger.html#dates)

`Date` is when the transaction occurred. It is usually written in the format YYYY/MM/DD. Ledgers support [secondary dates](https://hledger.org/hledger.html#secondary-dates), but it is recommended to stick with [posting dates](https://hledger.org/hledger.html#posting-dates).

This is a required component.

### [Transaction Comment](https://hledger.org/hledger.html#account-comments)

`Transaction Comment` provides additional information or notes for the transaction. It can be used to add any extra details about the transaction that are not covered in the description. Note that multiple account comment lines can be added, but they end at the first posting.

This component is optional.
>>>>>>> UPDATED
</source>