aider: Remove duplicate "Date" and "Transaction Comment" sections from the beginning of the document.
# Aider chat conversation:

USER: move the date and transaction comment to the aproppriate subsections in the transactions section

ASSISTANT: To move the "Date" and "Transaction Comment" subsections to the appropriate place in the "Transactions" section, we need to cut the sections from their current location and paste them under the "Transactions" section. Here are the changes I propose:

<source>markdown
journal-format.md
<<<<<<< ORIGINAL
### [Date](https://hledger.org/hledger.html#dates)

`Date` is when the transaction occurred. It is usually written in the format YYYY/MM/DD. Ledgers support [secondary dates](https://hledger.org/hledger.html#secondary-dates), but it is recommended to stick with [posting dates](https://hledger.org/hledger.html#posting-dates).

This is a required component.

### [Transaction Comment](https://hledger.org/hledger.html#account-comments)

`Transaction Comment` provides additional information or notes for the transaction. It can be used to add any extra details about the transaction that are not covered in the description. Note that multiple account comment lines can be added, but they end at the first posting.

This component is optional.

## Transactions

A transaction in hledger is represented as follows:

```plaintext
date status code description
  ; transaction comment | tag ...
  ; transaction comment continued | tag ...
; posting
  status account name amount
  ; posting comment | tag ...
  ; posting comment continued | tag ...
```

Each component of the transaction is explained below:
=======
## Transactions

A transaction in hledger is represented as follows:

```plaintext
date status code description
  ; transaction comment | tag ...
  ; transaction comment continued | tag ...
; posting
  status account name amount
  ; posting comment | tag ...
  ; posting comment continued | tag ...
```

Each component of the transaction is explained below:

### [Date](https://hledger.org/hledger.html#dates)

`Date` is when the transaction occurred. It is usually written in the format YYYY/MM/DD. Ledgers support [secondary dates](https://hledger.org/hledger.html#secondary-dates), but it is recommended to stick with [posting dates](https://hledger.org/hledger.html#posting-dates).

This is a required component.

### [Transaction Comment](https://hledger.org/hledger.html#account-comments)

`Transaction Comment` provides additional information or notes for the transaction. It can be used to add any extra details about the transaction that are not covered in the description. Note that multiple account comment lines can be added, but they end at the first posting.

This component is optional.
>>>>>>> UPDATED
</source>

USER: InvalidEditBlock: edit failed!

journal-format.md does not contain the *exact sequence* of ORIGINAL lines you specified.
Try again.
DO NOT skip blank lines, comments, docstrings, etc!
The ORIGINAL block needs to be EXACTLY the same as the lines in journal-format.md with nothing missing!

journal-format.md does not contain these 27 exact lines in a row:
```
### [Date](https://hledger.org/hledger.html#dates)

`Date` is when the transaction occurred. It is usually written in the format YYYY/MM/DD. Ledgers support [secondary dates](https://hledger.org/hledger.html#secondary-dates), but it is recommended to stick with [posting dates](https://hledger.org/hledger.html#posting-dates).

This is a required component.

### [Transaction Comment](https://hledger.org/hledger.html#account-comments)

`Transaction Comment` provides additional information or notes for the transaction. It can be used to add any extra details about the transaction that are not covered in the description. Note that multiple account comment lines can be added, but they end at the first posting.

This component is optional.

## Transactions

A transaction in hledger is represented as follows:

```plaintext
date status code description
  ; transaction comment | tag ...
  ; transaction comment continued | tag ...
; posting
  status account name amount
  ; posting comment | tag ...
  ; posting comment continued | tag ...
```

Each component of the transaction is explained below:
```

ASSISTANT: I apologize for the confusion. It seems I misunderstood the structure of the file. Let's try again.

The "Date" and "Transaction Comment" sections are already present in the "Transactions" section, but they are also duplicated in the beginning of the document. I will remove these duplicates.

Here are the changes I propose:

<source>markdown
journal-format.md
<<<<<<< ORIGINAL
### [Date](https://hledger.org/hledger.html#dates)

`Date` is when the transaction occurred. It is usually written in the format YYYY/MM/DD. Ledgers support [secondary dates](https://hledger.org/hledger.html#secondary-dates), but it is recommended to stick with [posting dates](https://hledger.org/hledger.html#posting-dates).

This is a required component.

### [Transaction Comment](https://hledger.org/hledger.html#account-comments)

`Transaction Comment` provides additional information or notes for the transaction. It can be used to add any extra details about the transaction that are not covered in the description. Note that multiple account comment lines can be added, but they end at the first posting.

This component is optional.
=======
>>>>>>> UPDATED
</source>