aider: Move "Date" and "Transaction Comment" to their respective sections in the transactions section.
# Aider chat conversation:

USER: Those weren't duplicates, they were additional data that needed to be moved to their sections in the transactions section.

ASSISTANT: I see, I misunderstood your request. I apologize for the confusion. You want to move the additional details about "Date" and "Transaction Comment" to their respective sections under "Transactions".

Here are the changes I propose:

<source>markdown
journal-format.md
<<<<<<< ORIGINAL
### [Date](https://hledger.org/hledger.html#dates)

`Date` is when the transaction occurred. It is usually written in the format YYYY/MM/DD.

This is a required component.
=======
### [Date](https://hledger.org/hledger.html#dates)

`Date` is when the transaction occurred. It is usually written in the format YYYY/MM/DD. Ledgers support [secondary dates](https://hledger.org/hledger.html#secondary-dates), but it is recommended to stick with [posting dates](https://hledger.org/hledger.html#posting-dates).

This is a required component.
>>>>>>> UPDATED
</source>

<source>markdown
journal-format.md
<<<<<<< ORIGINAL
### [Transaction Comment](https://hledger.org/hledger.html#account-comments)

`Transaction Comment` provides additional information or notes for the
transaction. It can be used to add any extra details about the transaction
that are not covered in the description.

This component is optional.
=======
### [Transaction Comment](https://hledger.org/hledger.html#account-comments)

`Transaction Comment` provides additional information or notes for the transaction. It can be used to add any extra details about the transaction that are not covered in the description. Note that multiple account comment lines can be added, but they end at the first posting.

This component is optional.
>>>>>>> UPDATED
</source>