aider: Move the links from the section headers to the first use of the concept name in each section.
# Aider chat conversation:

USER: Move the links from the section header to the first use of the concept name in that section.

ASSISTANT: To move the links from the section headers to the first use of the concept name in each section, we need to modify the `journal-format.md` file.

Here are the changes:

<source>markdown
journal-format.md
<<<<<<< ORIGINAL
### [Date](https://hledger.org/hledger.html#dates)

`Date` is when the transaction occurred. It is usually written in the format
YYYY/MM/DD.
=======
### Date

[`Date`](https://hledger.org/hledger.html#dates) is when the transaction occurred. It is usually written in the format
YYYY/MM/DD.
>>>>>>> UPDATED
</source>

<source>markdown
journal-format.md
<<<<<<< ORIGINAL
### [Status](https://hledger.org/hledger.html#status)

`Status` is the state of the transaction. It can be null, !, or \*. The status
is used to indicate if the transaction is pending (!) or cleared (\*).
=======
### Status

[`Status`](https://hledger.org/hledger.html#status) is the state of the transaction. It can be null, !, or \*. The status
is used to indicate if the transaction is pending (!) or cleared (\*).
>>>>>>> UPDATED
</source>

<source>markdown
journal-format.md
<<<<<<< ORIGINAL
### [Code](https://hledger.org/hledger.html#code)

`Code` is a reference number or identifier for the transaction. It is often
used to record check numbers or other reference information.
=======
### Code

[`Code`](https://hledger.org/hledger.html#code) is a reference number or identifier for the transaction. It is often
used to record check numbers or other reference information.
>>>>>>> UPDATED
</source>

<source>markdown
journal-format.md
<<<<<<< ORIGINAL
### [Description](https://hledger.org/hledger.html#description)

`Description` provides more details about the transaction. It is a brief
summary of the transaction's purpose.
=======
### Description

[`Description`](https://hledger.org/hledger.html#description) provides more details about the transaction. It is a brief
summary of the transaction's purpose.
>>>>>>> UPDATED
</source>

<source>markdown
journal-format.md
<<<<<<< ORIGINAL
### [Transaction Comment](https://hledger.org/1.30/hledger.html#transaction-comments)

`Transaction Comment` provides additional information or notes for the
transaction. It can be used to add any extra details about the transaction
that are not covered in the description. Note that multiple account comment
lines can be added, but they end at the first posting. Transaction comments
must be indented to be associated with the transaction.
=======
### Transaction Comment

[`Transaction Comment`](https://hledger.org/1.30/hledger.html#transaction-comments) provides additional information or notes for the
transaction. It can be used to add any extra details about the transaction
that are not covered in the description. Note that multiple account comment
lines can be added, but they end at the first posting. Transaction comments
must be indented to be associated with the transaction.
>>>>>>> UPDATED
</source>

<source>markdown
journal-format.md
<<<<<<< ORIGINAL
### [Tag](https://hledger.org/hledger.html#tags-1)

`Tag` is a keyword or label that helps in categorizing transactions. Tags can
be used to group related transactions together for reporting purposes.
=======
### Tag

[`Tag`](https://hledger.org/hledger.html#tags-1) is a keyword or label that helps in categorizing transactions. Tags can
be used to group related transactions together for reporting purposes.
>>>>>>> UPDATED
</source>

<source>markdown
journal-format.md
<<<<<<< ORIGINAL
### [Posting](https://hledger.org/hledger.html#virtual-postings)

`Posting` records the changes in account balances due to the transaction. Each
transaction must have at least one posting, and each posting affects the
balance of an account. The order of postings in a transaction can affect the
balance calculations, so it's important to maintain the correct order.
=======
### Posting

[`Posting`](https://hledger.org/hledger.html#virtual-postings) records the changes in account balances due to the transaction. Each
transaction must have at least one posting, and each posting affects the
balance of an account. The order of postings in a transaction can affect the
balance calculations, so it's important to maintain the correct order.
>>>>>>> UPDATED
</source>

<source>markdown
journal-format.md
<<<<<<< ORIGINAL
### [Account Name](https://hledger.org/hledger.html#account-names)

`Account Name` is the name of the account affected by the transaction. Account
names are hierarchical, and different levels of the hierarchy are separated by
colons.
=======
### Account Name

[`Account Name`](https://hledger.org/hledger.html#account-names) is the name of the account affected by the transaction. Account
names are hierarchical, and different levels of the hierarchy are separated by
colons.
>>>>>>> UPDATED
</source>

<source>markdown
journal-format.md
<<<<<<< ORIGINAL
### [Amount](https://hledger.org/hledger.html#amounts)

`Amount` represents the value of the transaction. Amounts can be in any
currency or commodity, and hledger will keep track of each one separately.
=======
### Amount

[`Amount`](https://hledger.org/hledger.html#amounts) represents the value of the transaction. Amounts can be in any
currency or commodity, and hledger will keep track of each one separately.
>>>>>>> UPDATED
</source>

<source>markdown
journal-format.md
<<<<<<< ORIGINAL
### [Posting Comment](https://hledger.org/hledger.html#comments)

`Posting Comment` provides additional information or notes about the posting.
It can be used to add any extra details about the posting that are not covered
in the account name or amount. Posting comments must be indented to be
associated with the posting.
=======
### Posting Comment

[`Posting Comment`](https://hledger.org/hledger.html#comments) provides additional information or notes about the posting.
It can be used to add any extra details about the posting that are not covered
in the account name or amount. Posting comments must be indented to be
associated with the posting.
>>>>>>> UPDATED
</source>

Please review these changes and let me know if they are correct.